Change MeController to ProfileController and add methods
# validate update
# update
